record: If message can't be encoded, fall back to latin-1 encoding
utf-8 is the correct encoding for the vast majority of messages, and we want to
represent them in the records as such for ease of use. However, some messages can't be
represented in utf-8 as-is. In these cases we fall back to 'decoding' into latin-1,
which will produce incorrect non-ascii characters but all byte strings are encodable.
We leave an extra flag to say we've done this.